backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.28
alembic>=1.13.1
psycopg[binary]>=3.1.18
pydantic[email]>=2.6.3
pydantic-settings>=2.2.1
passlib>=1.7.4
bcrypt==3.2.0
PyJWT>=2.8.0
python-multipart>=0.0.9
cryptography>=42.0.5
ffmpeg-python>=0.2.0
Pillow>=10.2.0
python-dotenv>=1.0.1
httpx>=0.27.0
fastapi-mail>=1.4.1
